+++ modules/common/src/print_util.c
@@ -12,10 +12,10 @@
 #include <stdint.h>
 #include <stdio.h>
 
-void print_array(const uint8_t *in_data, uint16_t in_len)
+void print_array(const uint8_t *in_data, uint32_t in_len)
 {
 	printf(" (size %u):", in_len);
-	for (uint16_t i = 0; i < in_len; i++) {
+	for (uint32_t i = 0; i < in_len; i++) {
 		if (i % 16 == 0)
 			printf("\n\t%02X ", in_data[i]);
 		else
