+++ modules/edhoc/inc/retrieve_cred.h
@@ -46,10 +46,10 @@ enum id_cred_x_label {
  * @param   g static DH public key
  * @param   g_len length of g
  */
-enum err retrieve_cred(bool static_dh_auth,
-			 struct other_party_cred *cred_array, uint16_t cred_num,
-			 uint8_t *id_cred, uint8_t id_cred_len, uint8_t *cred,
-			 uint16_t *cred_len, uint8_t *pk, uint16_t *pk_len,
-			 uint8_t *g, uint16_t *g_len);
+enum err retrieve_cred(bool static_dh_auth, struct other_party_cred *cred_array,
+		       uint16_t cred_num, uint8_t *id_cred,
+		       uint32_t id_cred_len, uint8_t *cred, uint32_t *cred_len,
+		       uint8_t *pk, uint32_t *pk_len, uint8_t *g,
+		       uint32_t *g_len);
 
 #endif
