+++ modules/edhoc/src/plaintext_decode.c
@@ -30,7 +30,7 @@
  * @param	
  */
 static enum err id_cred_x_encode(enum id_cred_x_label label, int algo,
-				 const void *id, uint64_t id_len,
+				 const void *id, uint32_t id_len,
 				 uint8_t *id_cred_x, uint32_t *id_cred_x_len)
 {
 	struct id_cred_x_map map;
@@ -71,7 +71,7 @@ static enum err id_cred_x_encode(enum id_cred_x_label label, int algo,
 	return ok;
 }
 
-enum err plaintext_split(uint8_t *ptxt, const uint16_t ptxt_len,
+enum err plaintext_split(uint8_t *ptxt, const uint32_t ptxt_len,
 			 uint8_t *id_cred_x, uint32_t *id_cred_x_len,
 			 uint8_t *sign_or_mac, uint32_t *sign_or_mac_len,
 			 uint8_t *ad, uint32_t *ad_len)
